Added More Paths
Added new paths and removed an old one
