+++ src/main/java/frc/robot/autonomous/Chooser.java
@@ -12,8 +12,8 @@ import frc.lib5k.utils.RobotLogger;
 import frc.robot.autonomous.actions.LogCommand;
 import frc.robot.autonomous.paths.AutonomousPath;
 import frc.robot.autonomous.paths.score.ScoreCenter;
+import frc.robot.autonomous.paths.score.ScoreGetBalls;
 import frc.robot.autonomous.paths.score.ScoreRight;
-import frc.robot.autonomous.paths.slow.SlowRight;
 import frc.robot.autonomous.paths.slow.SlowRightHalffield;
 import frc.robot.autonomous.paths.score.ScorePickupRight;
 import frc.robot.autonomous.paths.test.TestReverse;
@@ -38,13 +38,13 @@ public class Chooser {
     public Chooser() {
 
         // Paths
-        m_pathChooser.setDefaultOption("Right side slow (half-field shot) Test", new ScorePickupRight());
+        m_pathChooser.setDefaultOption("Score Get Balls", new ScoreGetBalls());
         m_pathChooser.addOption("Backwards test", new TestReverse());
         m_pathChooser.addOption("Test PID", new TestPID());
         m_pathChooser.addOption("Right side slow (half-field shot)", new SlowRightHalffield());
-        m_pathChooser.addOption("Right side slow", new SlowRight());
-        m_pathChooser.addOption("Score Right", new ScoreRight());
+        m_pathChooser.addOption("Right side slow (half-field shot) Test", new ScorePickupRight());
         m_pathChooser.addOption("Center Score", new ScoreCenter());
+        m_pathChooser.addOption("Score Right", new ScoreRight());
 
         // Scoring
         m_shouldScore.setDefaultOption("Score balls", true);
